pyproject.toml
[tool.poetry]
name = "goosebit"
version = "0.1.1"
description = ""
authors = ["Upstream Data <brett@upstreamdata.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["uvicorn"], version = "^0.111.0"}
python-multipart = "^0.0.9"
jinja2 = "^3.1.4"
itsdangerous = "^2.2.0"
tortoise-orm = "^0.21.4"
aerich = "^0.7.2"
aiofiles = "^24.1.0"
websockets = "^12.0"
argon2-cffi = "^23.1.0"
joserfc = "^1.0.0"
semver = "^3.0.2"
opentelemetry-distro = "^0.46b0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-exporter-prometheus = "^0.46b0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.2.0"
pre-commit = "^3.6.2"
pytest = "^8.3.1"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.29"

[tool.aerich]
tortoise_orm = "goosebit.db.TORTOISE_CONF"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
